moved edit name field out and changed update edit name function
