+++ app/src/state/parse-settings.ts
@@ -126,13 +126,18 @@ function processFonts(fontsNode: Element): FontsInfo {
   for (let i = 0; i < childNodes.length; i++) {
     const maybeChild = childNodes[i];
     if (maybeChild.nodeType === Node.ELEMENT_NODE) {
-      const child = <Element> maybeChild;
+      const child = <Element>maybeChild;
 
       const fontProps = child.textContent!.trim().split(";");
       for (let i = 0; i < fontProps.length; i++) {
-        fonts[child.getAttribute("function")! as keyof FontsInfo].inheritable = Boolean(child.getAttribute("inheritable"));
+        fonts[child.getAttribute("function")! as keyof FontsInfo].inheritable = Boolean(
+          child.getAttribute("inheritable")
+        );
         if (fontProps[i].includes("font-family")) {
-          fonts[child.getAttribute("function")! as keyof FontsInfo].fontFace = fontProps[i].split(": ")[1] === "Arial" || "Courier" || "Verdana" ? <"Arial" | "Courier" | "Verdana">fontProps[i].split(": ")[1] : "Arial";
+          fonts[child.getAttribute("function")! as keyof FontsInfo].fontFace =
+            fontProps[i].split(": ")[1] === "Arial" || "Courier" || "Verdana"
+              ? <"Arial" | "Courier" | "Verdana">fontProps[i].split(": ")[1]
+              : "Arial";
         }
         if (fontProps[i].includes("font-size")) {
           //ex 14px
